68000: Rework interrupt handling [O. Galibert]
* Implement the cpu space as an address space
* Make all vectored interrupts use the cpu space
* Make it possible to direct the cpu space to another space, use it for amiga (which handles it as a normal AS_PROGRAM read)
* Make it possible to disable the priority muxer and get 3 lines instead, use it for cps2

+++ src/devices/machine/mc68901.h
@@ -75,7 +75,7 @@ public:
 	DECLARE_READ8_MEMBER( read );
 	DECLARE_WRITE8_MEMBER( write );
 
-	int get_vector();
+	u16 get_vector();
 
 	DECLARE_WRITE_LINE_MEMBER( i0_w );
 	DECLARE_WRITE_LINE_MEMBER( i1_w );
